Move keymaps from handlers.lua to keymaps.lua.
